more fixes for updating to juce 6.1.5
