+++ modules/juce_audio_formats/codecs/juce_AudioFormatFileHelpers.h
@@ -2,7 +2,25 @@
 
 namespace juce
 {
+
+//==============================================================================
+using StringMap = std::unordered_map<String, String>;
+
+static auto toMap (const StringPairArray& array)
+{
+    StringMap result;
     
+    for (auto i = 0; i < array.size(); ++i)
+        result[array.getAllKeys()[i]] = array.getAllValues()[i];
+    
+    return result;
+}
+
+static auto getValueWithDefault (const StringMap& m, const String& key, const String& fallback = {})
+{
+    const auto iter = m.find (key);
+    return iter != m.cend() ? iter->second : fallback;
+}
 
 static const char* const wavFormatName = "WAV file";
 
@@ -134,13 +152,12 @@ const char* const WavAudioFormat::tracktionLoopInfo    = "tracktion loop info";
 //==============================================================================
 namespace WavFileHelpers
 {
-    JUCE_CONSTEXPR inline int chunkName (const char* name) noexcept         { return (int) ByteOrder::littleEndianInt (name); }
-    JUCE_CONSTEXPR inline size_t roundUpSize (size_t sz) noexcept           { return (sz + 3) & ~3u; }
+    constexpr inline int chunkName (const char* name) noexcept         { return (int) ByteOrder::littleEndianInt (name); }
+    constexpr inline size_t roundUpSize (size_t sz) noexcept           { return (sz + 3) & ~3u; }
     
 #if JUCE_MSVC
 #pragma pack (push, 1)
 #endif
-    
     struct iXMLChunk
     {
         String xmlString;
@@ -184,43 +201,42 @@ namespace WavFileHelpers
         uint8 reserved[190];
         char codingHistory[1];
         
-        void copyTo (StringPairArray& values, const int totalSize) const
+        void copyTo (StringMap& values, const int totalSize) const
         {
-            values.set (WavAudioFormat::bwavDescription,     String::fromUTF8 (description,     sizeof (description)));
-            values.set (WavAudioFormat::bwavOriginator,      String::fromUTF8 (originator,      sizeof (originator)));
-            values.set (WavAudioFormat::bwavOriginatorRef,   String::fromUTF8 (originatorRef,   sizeof (originatorRef)));
-            values.set (WavAudioFormat::bwavOriginationDate, String::fromUTF8 (originationDate, sizeof (originationDate)));
-            values.set (WavAudioFormat::bwavOriginationTime, String::fromUTF8 (originationTime, sizeof (originationTime)));
+            values[WavAudioFormat::bwavDescription]     = String::fromUTF8 (description,     sizeof (description));
+            values[WavAudioFormat::bwavOriginator]      = String::fromUTF8 (originator,      sizeof (originator));
+            values[WavAudioFormat::bwavOriginatorRef]   = String::fromUTF8 (originatorRef,   sizeof (originatorRef));
+            values[WavAudioFormat::bwavOriginationDate] = String::fromUTF8 (originationDate, sizeof (originationDate));
+            values[WavAudioFormat::bwavOriginationTime] = String::fromUTF8 (originationTime, sizeof (originationTime));
             
             auto timeLow  = ByteOrder::swapIfBigEndian (timeRefLow);
             auto timeHigh = ByteOrder::swapIfBigEndian (timeRefHigh);
             auto time = (((int64) timeHigh) << 32) + timeLow;
             
-            values.set (WavAudioFormat::bwavTimeReference, String (time));
-            values.set (WavAudioFormat::bwavCodingHistory,
-                        String::fromUTF8 (codingHistory, totalSize - (int) offsetof (BWAVChunk, codingHistory)));
+            values[WavAudioFormat::bwavTimeReference] = String (time);
+            values[WavAudioFormat::bwavCodingHistory] = String::fromUTF8 (codingHistory, totalSize - (int) offsetof (BWAVChunk, codingHistory));
         }
         
-        static MemoryBlock createFrom (const StringPairArray& values)
+        static MemoryBlock createFrom (const StringMap& values)
         {
-            MemoryBlock data (roundUpSize (sizeof (BWAVChunk) + values[WavAudioFormat::bwavCodingHistory].getNumBytesAsUTF8()));
+            MemoryBlock data (roundUpSize (sizeof (BWAVChunk) + getValueWithDefault (values, WavAudioFormat::bwavCodingHistory).getNumBytesAsUTF8()));
             data.fillWith (0);
             
             auto* b = (BWAVChunk*) data.getData();
             
             // Allow these calls to overwrite an extra byte at the end, which is fine as long
-            // as they get called in the right order..
-            values[WavAudioFormat::bwavDescription]    .copyToUTF8 (b->description, 257);
-            values[WavAudioFormat::bwavOriginator]     .copyToUTF8 (b->originator, 33);
-            values[WavAudioFormat::bwavOriginatorRef]  .copyToUTF8 (b->originatorRef, 33);
-            values[WavAudioFormat::bwavOriginationDate].copyToUTF8 (b->originationDate, 11);
-            values[WavAudioFormat::bwavOriginationTime].copyToUTF8 (b->originationTime, 9);
+            // as they get called in the right order.
+            getValueWithDefault (values, WavAudioFormat::bwavDescription)    .copyToUTF8 (b->description, 257);
+            getValueWithDefault (values, WavAudioFormat::bwavOriginator)     .copyToUTF8 (b->originator, 33);
+            getValueWithDefault (values, WavAudioFormat::bwavOriginatorRef)  .copyToUTF8 (b->originatorRef, 33);
+            getValueWithDefault (values, WavAudioFormat::bwavOriginationDate).copyToUTF8 (b->originationDate, 11);
+            getValueWithDefault (values, WavAudioFormat::bwavOriginationTime).copyToUTF8 (b->originationTime, 9);
             
-            auto time = values[WavAudioFormat::bwavTimeReference].getLargeIntValue();
+            auto time = getValueWithDefault (values, WavAudioFormat::bwavTimeReference).getLargeIntValue();
             b->timeRefLow = ByteOrder::swapIfBigEndian ((uint32) (time & 0xffffffff));
             b->timeRefHigh = ByteOrder::swapIfBigEndian ((uint32) (time >> 32));
             
-            values[WavAudioFormat::bwavCodingHistory].copyToUTF8 (b->codingHistory, 0x7fffffff);
+            getValueWithDefault (values, WavAudioFormat::bwavCodingHistory).copyToUTF8 (b->codingHistory, 0x7fffffff);
             
             if (b->description[0] != 0
                 || b->originator[0] != 0
@@ -277,17 +293,17 @@ namespace WavFileHelpers
         SampleLoop loops[1];
         
         template <typename NameType>
-        static void setValue (StringPairArray& values, NameType name, uint32 val)
+        static void setValue (StringMap& values, NameType name, uint32 val)
         {
-            values.set (name, String (ByteOrder::swapIfBigEndian (val)));
+            values[name] = String (ByteOrder::swapIfBigEndian (val));
         }
         
-        static void setValue (StringPairArray& values, int prefix, const char* name, uint32 val)
+        static void setValue (StringMap& values, int prefix, const char* name, uint32 val)
         {
             setValue (values, "Loop" + String (prefix) + name, val);
         }
         
-        void copyTo (StringPairArray& values, const int totalSize) const
+        void copyTo (StringMap& values, const int totalSize) const
         {
             setValue (values, "Manufacturer",      manufacturer);
             setValue (values, "Product",           product);
@@ -314,20 +330,20 @@ namespace WavFileHelpers
         }
         
         template <typename NameType>
-        static uint32 getValue (const StringPairArray& values, NameType name, const char* def)
+        static uint32 getValue (const StringMap& values, NameType name, const char* def)
         {
-            return ByteOrder::swapIfBigEndian ((uint32) values.getValue (name, def).getIntValue());
+            return ByteOrder::swapIfBigEndian ((uint32) getValueWithDefault (values, name, def).getIntValue());
         }
         
-        static uint32 getValue (const StringPairArray& values, int prefix, const char* name, const char* def)
+        static uint32 getValue (const StringMap& values, int prefix, const char* name, const char* def)
         {
             return getValue (values, "Loop" + String (prefix) + name, def);
         }
         
-        static MemoryBlock createFrom (const StringPairArray& values)
+        static MemoryBlock createFrom (const StringMap& values)
         {
             MemoryBlock data;
-            auto numLoops = jmin (64, values.getValue ("NumSampleLoops", "0").getIntValue());
+            auto numLoops = jmin (64, getValueWithDefault (values, "NumSampleLoops", "0").getIntValue());
             
             data.setSize (roundUpSize (sizeof (SMPLChunk) + (size_t) (jmax (0, numLoops - 1)) * sizeof (SampleLoop)), true);
             
@@ -369,12 +385,12 @@ namespace WavFileHelpers
         int8 lowVelocity;
         int8 highVelocity;
         
-        static void setValue (StringPairArray& values, const char* name, int val)
+        static void setValue (StringMap& values, const char* name, int val)
         {
-            values.set (name, String (val));
+            values[name] = String (val);
         }
         
-        void copyTo (StringPairArray& values) const
+        void copyTo (StringMap& values) const
         {
             setValue (values, "MidiUnityNote",  baseNote);
             setValue (values, "Detune",         detune);
@@ -385,17 +401,17 @@ namespace WavFileHelpers
             setValue (values, "HighVelocity",   highVelocity);
         }
         
-        static int8 getValue (const StringPairArray& values, const char* name, const char* def)
+        static int8 getValue (const StringMap& values, const char* name, const char* def)
         {
-            return (int8) values.getValue (name, def).getIntValue();
+            return (int8) getValueWithDefault (values, name, def).getIntValue();
         }
         
-        static MemoryBlock createFrom (const StringPairArray& values)
+        static MemoryBlock createFrom (const StringMap& values)
         {
             MemoryBlock data;
-            auto& keys = values.getAllKeys();
             
-            if (keys.contains ("LowNote", true) && keys.contains ("HighNote", true))
+            if (   values.find ("LowNote")  != values.cend()
+                && values.find ("HighNote") != values.cend())
             {
                 data.setSize (8, true);
                 auto* inst = static_cast<InstChunk*> (data.getData());
@@ -429,14 +445,14 @@ namespace WavFileHelpers
         uint32 numCues;
         Cue cues[1];
         
-        static void setValue (StringPairArray& values, int prefix, const char* name, uint32 val)
+        static void setValue (StringMap& values, int prefix, const char* name, uint32 val)
         {
-            values.set ("Cue" + String (prefix) + name, String (ByteOrder::swapIfBigEndian (val)));
+            values["Cue" + String (prefix) + name] = String (ByteOrder::swapIfBigEndian (val));
         }
         
-        void copyTo (StringPairArray& values, const int totalSize) const
+        void copyTo (StringMap& values, const int totalSize) const
         {
-            values.set ("NumCuePoints", String (ByteOrder::swapIfBigEndian (numCues)));
+            values["NumCuePoints"] = String (ByteOrder::swapIfBigEndian (numCues));
             
             for (int i = 0; i < (int) numCues; ++i)
             {
@@ -452,10 +468,10 @@ namespace WavFileHelpers
             }
         }
         
-        static MemoryBlock createFrom (const StringPairArray& values)
+        static MemoryBlock createFrom (const StringMap& values)
         {
             MemoryBlock data;
-            const int numCues = values.getValue ("NumCuePoints", "0").getIntValue();
+            const int numCues = getValueWithDefault (values, "NumCuePoints", "0").getIntValue();
             
             if (numCues > 0)
             {
@@ -475,23 +491,23 @@ namespace WavFileHelpers
                 for (int i = 0; i < numCues; ++i)
                 {
                     auto prefix = "Cue" + String (i);
-                    auto identifier = (uint32) values.getValue (prefix + "Identifier", "0").getIntValue();
+                    auto identifier = (uint32) getValueWithDefault (values, prefix + "Identifier", "0").getIntValue();
                     
 #if JUCE_DEBUG
                     jassert (! identifiers.contains (identifier));
                     identifiers.add (identifier);
 #endif
                     
-                    auto order = values.getValue (prefix + "Order", String (nextOrder)).getIntValue();
+                    auto order = getValueWithDefault (values, prefix + "Order", String (nextOrder)).getIntValue();
                     nextOrder = jmax (nextOrder, order) + 1;
                     
                     auto& cue = c->cues[i];
                     cue.identifier   = ByteOrder::swapIfBigEndian ((uint32) identifier);
                     cue.order        = ByteOrder::swapIfBigEndian ((uint32) order);
-                    cue.chunkID      = ByteOrder::swapIfBigEndian ((uint32) values.getValue (prefix + "ChunkID", dataChunkID).getIntValue());
-                    cue.chunkStart   = ByteOrder::swapIfBigEndian ((uint32) values.getValue (prefix + "ChunkStart", "0").getIntValue());
-                    cue.blockStart   = ByteOrder::swapIfBigEndian ((uint32) values.getValue (prefix + "BlockStart", "0").getIntValue());
-                    cue.offset       = ByteOrder::swapIfBigEndian ((uint32) values.getValue (prefix + "Offset", "0").getIntValue());
+                    cue.chunkID      = ByteOrder::swapIfBigEndian ((uint32) getValueWithDefault (values, prefix + "ChunkID", dataChunkID).getIntValue());
+                    cue.chunkStart   = ByteOrder::swapIfBigEndian ((uint32) getValueWithDefault (values, prefix + "ChunkStart", "0").getIntValue());
+                    cue.blockStart   = ByteOrder::swapIfBigEndian ((uint32) getValueWithDefault (values, prefix + "BlockStart", "0").getIntValue());
+                    cue.offset       = ByteOrder::swapIfBigEndian ((uint32) getValueWithDefault (values, prefix + "Offset", "0").getIntValue());
                 }
             }
             
@@ -503,20 +519,20 @@ namespace WavFileHelpers
     //==============================================================================
     namespace ListChunk
     {
-        static int getValue (const StringPairArray& values, const String& name)
+        static int getValue (const StringMap& values, const String& name)
         {
-            return values.getValue (name, "0").getIntValue();
+            return getValueWithDefault (values, name, "0").getIntValue();
         }
         
-        static int getValue (const StringPairArray& values, const String& prefix, const char* name)
+        static int getValue (const StringMap& values, const String& prefix, const char* name)
         {
             return getValue (values, prefix + name);
         }
         
-        static void appendLabelOrNoteChunk (const StringPairArray& values, const String& prefix,
+        static void appendLabelOrNoteChunk (const StringMap& values, const String& prefix,
                                             const int chunkType, MemoryOutputStream& out)
         {
-            auto label = values.getValue (prefix + "Text", prefix);
+            auto label = getValueWithDefault (values, prefix + "Text", prefix);
             auto labelLength = (int) label.getNumBytesAsUTF8() + 1;
             auto chunkLength = 4 + labelLength + (labelLength & 1);
             
@@ -529,9 +545,9 @@ namespace WavFileHelpers
                 out.writeByte (0);
         }
         
-        static void appendExtraChunk (const StringPairArray& values, const String& prefix, MemoryOutputStream& out)
+        static void appendExtraChunk (const StringMap& values, const String& prefix, MemoryOutputStream& out)
         {
-            auto text = values.getValue (prefix + "Text", prefix);
+            auto text = getValueWithDefault (values, prefix + "Text", prefix);
             
             auto textLength = (int) text.getNumBytesAsUTF8() + 1; // include null terminator
             auto chunkLength = textLength + 20 + (textLength & 1);
@@ -551,7 +567,7 @@ namespace WavFileHelpers
                 out.writeByte (0);
         }
         
-        static MemoryBlock createFrom (const StringPairArray& values)
+        static MemoryBlock createFrom (const StringMap& values)
         {
             auto numCueLabels  = getValue (values, "NumCueLabels");
             auto numCueNotes   = getValue (values, "NumCueNotes");
@@ -675,7 +691,7 @@ namespace WavFileHelpers
             return true;
         }
         
-        static void addToMetadata (StringPairArray& values, InputStream& input, int64 chunkEnd)
+        static void addToMetadata (StringMap& values, InputStream& input, int64 chunkEnd)
         {
             while (input.getPosition() < chunkEnd)
             {
@@ -695,8 +711,8 @@ namespace WavFileHelpers
                         {
                             MemoryBlock mb;
                             input.readIntoMemoryBlock (mb, (ssize_t) infoLength);
-                            values.set (type, String::createStringFromData ((const char*) mb.getData(),
-                                                                            (int) mb.getSize()));
+                            values[type] = String::createStringFromData ((const char*) mb.getData(),
+                                                                         (int) mb.getSize());
                             break;
                         }
                     }
@@ -704,9 +720,9 @@ namespace WavFileHelpers
             }
         }
         
-        static bool writeValue (const StringPairArray& values, MemoryOutputStream& out, const char* paramName)
+        static bool writeValue (const StringMap& values, MemoryOutputStream& out, const char* paramName)
         {
-            auto value = values.getValue (paramName, {});
+            auto value = getValueWithDefault (values, paramName, {});
             
             if (value.isEmpty())
                 return false;
@@ -724,7 +740,7 @@ namespace WavFileHelpers
             return true;
         }
         
-        static MemoryBlock createFrom (const StringPairArray& values)
+        static MemoryBlock createFrom (const StringMap& values)
         {
             MemoryOutputStream out;
             out.writeInt (chunkName ("INFO"));
@@ -748,7 +764,7 @@ namespace WavFileHelpers
             input.read (this, (int) jmin (sizeof (*this), length));
         }
         
-        AcidChunk (const StringPairArray& values)
+        AcidChunk (const StringMap& values)
         {
             zerostruct (*this);
             
@@ -758,18 +774,20 @@ namespace WavFileHelpers
             | getFlagIfPresent (values, WavAudioFormat::acidDiskBased, 0x08)
             | getFlagIfPresent (values, WavAudioFormat::acidizerFlag,  0x10);
             
-            if (values[WavAudioFormat::acidRootSet].getIntValue() != 0)
-                rootNote = ByteOrder::swapIfBigEndian ((uint16) values[WavAudioFormat::acidRootNote].getIntValue());
+            if (getValueWithDefault (values, WavAudioFormat::acidRootSet).getIntValue() != 0)
+                rootNote = ByteOrder::swapIfBigEndian ((uint16) getValueWithDefault (values, WavAudioFormat::acidRootNote).getIntValue());
             
-            numBeats          = ByteOrder::swapIfBigEndian ((uint32) values[WavAudioFormat::acidBeats].getIntValue());
-            meterDenominator  = ByteOrder::swapIfBigEndian ((uint16) values[WavAudioFormat::acidDenominator].getIntValue());
-            meterNumerator    = ByteOrder::swapIfBigEndian ((uint16) values[WavAudioFormat::acidNumerator].getIntValue());
+            numBeats          = ByteOrder::swapIfBigEndian ((uint32) getValueWithDefault (values, WavAudioFormat::acidBeats).getIntValue());
+            meterDenominator  = ByteOrder::swapIfBigEndian ((uint16) getValueWithDefault (values, WavAudioFormat::acidDenominator).getIntValue());
+            meterNumerator    = ByteOrder::swapIfBigEndian ((uint16) getValueWithDefault (values, WavAudioFormat::acidNumerator).getIntValue());
             
-            if (values.containsKey (WavAudioFormat::acidTempo))
-                tempo = swapFloatByteOrder (values[WavAudioFormat::acidTempo].getFloatValue());
+            const auto iter = values.find (WavAudioFormat::acidTempo);
+            
+            if (iter != values.cend())
+                tempo = swapFloatByteOrder (iter->second.getFloatValue());
         }
         
-        static MemoryBlock createFrom (const StringPairArray& values)
+        static MemoryBlock createFrom (const StringMap& values)
         {
             return AcidChunk (values).toMemoryBlock();
         }
@@ -780,7 +798,7 @@ namespace WavFileHelpers
             ? MemoryBlock (this, sizeof (*this)) : MemoryBlock();
         }
         
-        void addToMetadata (StringPairArray& values) const
+        void addToMetadata (StringMap& values) const
         {
             setBoolFlag (values, WavAudioFormat::acidOneShot,   0x01);
             setBoolFlag (values, WavAudioFormat::acidRootSet,   0x02);
@@ -789,22 +807,22 @@ namespace WavFileHelpers
             setBoolFlag (values, WavAudioFormat::acidizerFlag,  0x10);
             
             if (flags & 0x02) // root note set
-                values.set (WavAudioFormat::acidRootNote, String (ByteOrder::swapIfBigEndian (rootNote)));
+                values[WavAudioFormat::acidRootNote] = String (ByteOrder::swapIfBigEndian (rootNote));
             
-            values.set (WavAudioFormat::acidBeats,       String (ByteOrder::swapIfBigEndian (numBeats)));
-            values.set (WavAudioFormat::acidDenominator, String (ByteOrder::swapIfBigEndian (meterDenominator)));
-            values.set (WavAudioFormat::acidNumerator,   String (ByteOrder::swapIfBigEndian (meterNumerator)));
-            values.set (WavAudioFormat::acidTempo,       String (swapFloatByteOrder (tempo)));
+            values[WavAudioFormat::acidBeats]       = String (ByteOrder::swapIfBigEndian (numBeats));
+            values[WavAudioFormat::acidDenominator] = String (ByteOrder::swapIfBigEndian (meterDenominator));
+            values[WavAudioFormat::acidNumerator]   = String (ByteOrder::swapIfBigEndian (meterNumerator));
+            values[WavAudioFormat::acidTempo]       = String (swapFloatByteOrder (tempo));
         }
         
-        void setBoolFlag (StringPairArray& values, const char* name, uint32 mask) const
+        void setBoolFlag (StringMap& values, const char* name, uint32 mask) const
         {
-            values.set (name, (flags & ByteOrder::swapIfBigEndian (mask)) ? "1" : "0");
+            values[name] = (flags & ByteOrder::swapIfBigEndian (mask)) ? "1" : "0";
         }
         
-        static uint32 getFlagIfPresent (const StringPairArray& values, const char* name, uint32 flag)
+        static uint32 getFlagIfPresent (const StringMap& values, const char* name, uint32 flag)
         {
-            return values[name].getIntValue() != 0 ? ByteOrder::swapIfBigEndian (flag) : 0;
+            return getValueWithDefault (values, name).getIntValue() != 0 ? ByteOrder::swapIfBigEndian (flag) : 0;
         }
         
         static float swapFloatByteOrder (const float x) noexcept
@@ -833,10 +851,10 @@ namespace WavFileHelpers
     //==============================================================================
     struct TracktionChunk
     {
-        static MemoryBlock createFrom (const StringPairArray& values)
+        static MemoryBlock createFrom (const StringMap& values)
         {
             MemoryOutputStream out;
-            auto s = values[WavAudioFormat::tracktionLoopInfo];
+            auto s = getValueWithDefault (values, WavAudioFormat::tracktionLoopInfo);
             
             if (s.isNotEmpty())
             {
@@ -853,7 +871,7 @@ namespace WavFileHelpers
     //==============================================================================
     namespace AXMLChunk
     {
-        static void addToMetadata (StringPairArray& destValues, const String& source)
+        static void addToMetadata (StringMap& destValues, const String& source)
         {
             if (auto xml = parseXML (source))
             {
@@ -868,7 +886,7 @@ namespace WavFileHelpers
                                 auto ISRCCode = xml4->getAllSubText().fromFirstOccurrenceOf ("ISRC:", false, true);
                                 
                                 if (ISRCCode.isNotEmpty())
-                                    destValues.set (WavAudioFormat::ISRC, ISRCCode);
+                                    destValues[WavAudioFormat::ISRC] = ISRCCode;
                             }
                         }
                     }
@@ -876,9 +894,9 @@ namespace WavFileHelpers
             }
         }
         
-        static MemoryBlock createFrom (const StringPairArray& values)
+        static MemoryBlock createFrom (const StringMap& values)
         {
-            auto ISRC = values.getValue (WavAudioFormat::ISRC, {});
+            auto ISRC = getValueWithDefault (values, WavAudioFormat::ISRC);
             MemoryOutputStream xml;
             
             if (ISRC.isNotEmpty())
@@ -930,7 +948,7 @@ namespace WavFileHelpers
         uint32 sampleCountHigh; // high 4 byte sample count of fact chunk
         uint32 tableLength;     // number of valid entries in array 'table'
     } JUCE_PACKED;
-    
+
 #if JUCE_MSVC
 #pragma pack (pop)
 #endif
